feat: Refactor to extract hooks and components
This commit introduces a major refactor of the codebase, extracting reusable hooks and components to improve code organization and readability.

Key changes:

    Extracted custom hooks:

        useFileHandlers: Manages file and folder operations, including loading and updating the list of items.

        useSearchHandlers: Handles search functionality, filtering items based on the search query.

    Extracted components:

        ItemRenderer: Renders the list of items, highlighting current and selected items.

        Navigation: Handles navigation between items in the list.

        Handlers: Manages user input for toggling folder expansion, item selection, and copying file contents.

        InputHandler: Handles user input, including search queries, navigation, and other actions.

    Introduced type definitions:

        Created types.ts to define the Item interface for better type safety.

    Added constants.ts:

        Defined constants for version numbers and excluded folders for better maintainability.

    Created syncVersion.mjs:

        Added a script to synchronize the version number between the package.json and the source code.

This refactor enhances code modularity, making it easier to understand, maintain, and extend the functionality of the application.
