+++ source/constants.ts
@@ -1,0 +1,7 @@
+// constants.ts
+
+// Define the version numbers.
+export const AIFORMAT_VERSION = '0.1.4'; // Incremented version number
+export const GENERATE_OUTPUT_VERSION = "0.0.4"; // Incremented version number
+// Define the default folders to be excluded.
+export const EXCLUDED_FOLDERS = ['node_modules', '.git', 'dist', 'build', 'coverage'];
